pyproject.toml
[tool.poetry]
name = "sp-allure-docker-cicd"
version = "0.1.0"
description = "Selenium with Allure reporting"
authors = ["edmeister <koed.dev@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.18.0"
pytest = "^8.1.1"
allure-pytest = "^2.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
